backend/requirements.txt
# FastAPI Backend Requirements
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.10.0
pydantic-settings==2.6.0
python-multipart==0.0.12
websockets==13.1
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
